Python/requirements.txt
absl-py==0.6.1
astor==0.7.1
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
imageio==2.4.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.14.5
Pillow==5.3.0
protobuf==3.6.1
scikit-learn==0.20.0
scipy==1.1.0
SimpleITK==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
torch==0.4.1
Werkzeug==0.14.1
